callback: truncate inventory hostnames longer than 255 characters
Documentation suggests (https://www.freesoft.org/CIE/RFC/1035/9.htm)
that DNS hostnames cannot be longer than 255 characters but there are no
such limits for Ansible inventory hostnames.

Trying to send an inventory hostname that is longer than 255 characters
to the API trips into an exception because the field is a varchar(255).

Truncate the hostname to prevent an exception and warn about it once per
host if it happens.