Add support for recording versions of ara
Recording versions of the ara client and server that were involved may
prove helpful for comparing recordings across versions and
troubleshooting issues.

This commit includes a database migration for adding the "client_version"
and "server_version" fields to the playbook model, API and CLI commands.

In a typical scenario, the callback plugin supplies it's version as
client_version and the server_version field is automatically populated
by the server.

The UI implementation is limited for this first iteration:
- It is possible to search for ?client_version and ?server_version
  parameters by manually specifying them in the playbook index URL:
  there are no search form fields for them
- The client_version and server_version are only shown in the playbook
  report card at the top of the page.

Drive-by improvements:
- Fix phrasing for --user help on the playbook CLI commands
- Set the version of ara once under ara.setup and use that everywhere
  instead of calling pbr.version every time
- Re-order columns in the playbook report card such that item counts are
  further to the right as they are less relevant in that context